requirements.txt
#pytorch
torchvision
efficientnet_pytorch
scikit-learn==0.24.1
numpy==1.19.5
scipy==1.5.4
pandas==1.1.5
joblib==1.0.1
Pillow==8.0.1
matplotlib==3.3.4
